Cloud/ARIMA/requirements.txt
pandas
matplotlib
statsmodels
kafka-python
kubernetes
requests
